Pygments-based Code Highlighting with Theme
Code themes for editors are now also integrated with Pygments-based
code formatting (as a result of exporting a notebook in HTML).

The code coloring CSS for Pygments tries to mimic the one in
standard notebook based on CodeMirror.
The alignment is not fuuully 100% compatible though due to
intrinsic differences in the two interpreters.

Additional colour support has been added for string2
(i.e. backtick strings)