All local fonts removed, and resorted to remote fonts via CDN
Now the theme is even more portable, allowing the use of Network remote fonts (via CDN) rather than local copies.

Due to lack of availability of Hasklig font on CDN, the default font for code editing is now changed to Fira Code.
Relying on remotely available fonts improves (by a ton) the portability of notebooks (especially for notebook-magic version of the theme).
On one hand, this allows to get rid of the serverextension - simplifying the whole thing.
Moreover, visualising the notebook on other computers **not having the theme** installed should not harm the style, as it will embedded in the notebook itself. Therefore, everything should be working, provided the notebook is trusted (so that custom HTML visualisation is enabled).