Completely removed dependency from local fonts
As per the Custom Jupyter theme, no local fonts are anymore needed. All default fonts are automatically downloaded from CDN.

This improves the portability of the theme within notebooks, without requiring any further installations from the user.
The only requirement is that notebook is **trusted** to enable HTML display. This is compulsory, and in general compliant with Jupyter notebook security protections.