Renewed documentation, targeted for PyPi package description.
The README on this branch now contains a clearer description of the package, and a lighter description of the whole projecsts with due credits and acks.

Other versions of this file (in other branches) will integrate part of this, and specialised only in the installation of the specific flavour.
Only the `master` branch will hold everything together into a single unique file.